refactor: consolidate DailyTasksRenderer test files
Consolidate 9 unit test files into 5 files for better maintainability:

Before (9 files):
- render.test.ts (14 tests)
- timestamps.test.ts (4 tests)
- extractFirstValue.test.ts (5 tests)
- getAssetLabel.test.ts (2 active tests)
- getEffortArea.test.ts (3 active tests)
- filtering.test.ts (4 tests)
- callbacks.test.ts (4 tests)
- edgeCases.test.ts (4 tests)
- properties.test.ts (6 tests)

After (5 files):
- core.test.ts (18 tests - merged render + timestamps)
- helpers.test.ts (10 tests - merged extractFirstValue + getAssetLabel + getEffortArea)
- callbacks.test.ts (4 tests - unchanged)
- edgeCases.test.ts (4 tests - unchanged)
- properties.test.ts (6 tests - unchanged)

Changes:
- Remove filtering.test.ts (covered by UI test in UniversalLayoutRenderer.ui.test.ts)
- Merge helper method tests into DailyTasksRenderer.helpers.test.ts
- Merge render + timestamps tests into DailyTasksRenderer.core.test.ts
- Total: 42 active tests (was 46, removed 4 duplicate filtering tests)

Closes #618
