+++ packages/obsidian-plugin/tests/unit/renderers/dailyTasks/DailyTasksRenderer.core.test.ts
@@ -5,6 +5,15 @@ import {
   DailyTasksRendererTestContext,
 } from "./DailyTasksRenderer.fixtures";
 
+/**
+ * Consolidated core tests for DailyTasksRenderer:
+ * - Main rendering logic
+ * - Timestamp formatting
+ *
+ * These tests verify the primary render() method behavior
+ * including file type detection, task rendering, and timestamp handling.
+ */
+
 describe("DailyTasksRenderer - render", () => {
   let ctx: DailyTasksRendererTestContext;
 
@@ -563,6 +572,192 @@ describe("DailyTasksRenderer - render", () => {
       ems__Effort_status: "[[ems__EffortStatusDoing]]",
       ems__Effort_startTimestamp: "2025-10-20T09:00:00.000Z",
       ems__Effort_endTimestamp: "2025-10-20T17:00:00.000Z",
+    };
+
+    ctx.mockMetadataExtractor.extractMetadata
+      .mockReturnValueOnce(metadata)
+      .mockReturnValueOnce(taskMetadata);
+
+    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
+      "[[pn__DailyNote]]",
+    );
+
+    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
+
+    const mockEl = createMockElement();
+    await ctx.renderer.render(mockEl, mockFile);
+
+    expect(ctx.mockReactRenderer.render).toHaveBeenCalled();
+    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
+    const tasks = renderCall[1].props.tasks;
+    expect(tasks[0].startTime).toBeTruthy();
+    expect(tasks[0].endTime).toBeTruthy();
+  });
+});
+
+describe("DailyTasksRenderer - timestamp formatting", () => {
+  let ctx: DailyTasksRendererTestContext;
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+    ctx = setupDailyTasksRendererTest();
+  });
+
+  it("should format valid timestamps correctly", async () => {
+    const mockFile = {
+      path: "test.md",
+      parent: { path: "DailyNotes" },
+      basename: "2025-10-20",
+    } as TFile;
+    const metadata = {
+      exo__Instance_class: "[[pn__DailyNote]]",
+      pn__DailyNote_day: "[[2025-10-20]]",
+    };
+
+    const taskFile = {
+      path: "task.md",
+      basename: "test-task",
+    } as TFile;
+
+    const taskMetadata = {
+      exo__Instance_class: "[[ems__Task]]",
+      ems__Effort_day: "[[2025-10-20]]",
+      ems__Effort_status: "[[ems__EffortStatusDoing]]",
+      ems__Effort_startTimestamp: "2025-10-20T09:00:00.000Z",
+      ems__Effort_endTimestamp: "2025-10-20T17:00:00.000Z",
+    };
+
+    ctx.mockMetadataExtractor.extractMetadata
+      .mockReturnValueOnce(metadata)
+      .mockReturnValueOnce(taskMetadata);
+
+    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
+      "[[pn__DailyNote]]",
+    );
+
+    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
+
+    const mockEl = createMockElement();
+    await ctx.renderer.render(mockEl, mockFile);
+
+    expect(ctx.mockReactRenderer.render).toHaveBeenCalled();
+    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
+    const tasks = renderCall[1].props.tasks;
+    expect(tasks[0].startTime).toBeTruthy();
+    expect(tasks[0].endTime).toBeTruthy();
+  });
+
+  it("should handle invalid timestamps gracefully", async () => {
+    const mockFile = {
+      path: "test.md",
+      parent: { path: "DailyNotes" },
+      basename: "2025-10-20",
+    } as TFile;
+    const metadata = {
+      exo__Instance_class: "[[pn__DailyNote]]",
+      pn__DailyNote_day: "[[2025-10-20]]",
+    };
+
+    const taskFile = {
+      path: "task.md",
+      basename: "test-task",
+    } as TFile;
+
+    const taskMetadata = {
+      exo__Instance_class: "[[ems__Task]]",
+      ems__Effort_day: "[[2025-10-20]]",
+      ems__Effort_plannedStartTimestamp: "2025-10-20T09:00:00",
+      ems__Effort_status: "[[ems__EffortStatusDoing]]",
+      ems__Effort_startTimestamp: "invalid",
+      ems__Effort_endTimestamp: "also-invalid",
+    };
+
+    ctx.mockMetadataExtractor.extractMetadata
+      .mockReturnValueOnce(metadata)
+      .mockReturnValueOnce(taskMetadata);
+
+    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
+      "[[pn__DailyNote]]",
+    );
+
+    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
+
+    const mockEl = createMockElement();
+    await ctx.renderer.render(mockEl, mockFile);
+
+    expect(ctx.mockReactRenderer.render).toHaveBeenCalled();
+    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
+    const tasks = renderCall[1].props.tasks;
+    expect(tasks[0].startTime).toBe("09:00");
+    expect(tasks[0].endTime).toBe("");
+  });
+
+  it("should handle null/undefined timestamps", async () => {
+    const mockFile = {
+      path: "test.md",
+      parent: { path: "DailyNotes" },
+      basename: "2025-10-20",
+    } as TFile;
+    const metadata = {
+      exo__Instance_class: "[[pn__DailyNote]]",
+      pn__DailyNote_day: "[[2025-10-20]]",
+    };
+
+    const taskFile = {
+      path: "task.md",
+      basename: "test-task",
+    } as TFile;
+
+    const taskMetadata = {
+      exo__Instance_class: "[[ems__Task]]",
+      ems__Effort_day: "[[2025-10-20]]",
+      ems__Effort_status: "[[ems__EffortStatusDoing]]",
+      ems__Effort_plannedEndTimestamp: "2025-10-20T18:00:00",
+    };
+
+    ctx.mockMetadataExtractor.extractMetadata
+      .mockReturnValueOnce(metadata)
+      .mockReturnValueOnce(taskMetadata);
+
+    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
+      "[[pn__DailyNote]]",
+    );
+
+    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
+
+    const mockEl = createMockElement();
+    await ctx.renderer.render(mockEl, mockFile);
+
+    expect(ctx.mockReactRenderer.render).toHaveBeenCalled();
+    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
+    const tasks = renderCall[1].props.tasks;
+    expect(tasks[0].startTime).toBe("");
+    expect(tasks[0].endTime).toBe("18:00");
+  });
+
+  it("should fallback to planned timestamps when actual not available", async () => {
+    const mockFile = {
+      path: "test.md",
+      parent: { path: "DailyNotes" },
+      basename: "2025-10-20",
+    } as TFile;
+    const metadata = {
+      exo__Instance_class: "[[pn__DailyNote]]",
+      pn__DailyNote_day: "[[2025-10-20]]",
+    };
+
+    const taskFile = {
+      path: "task.md",
+      basename: "test-task",
+    } as TFile;
+
+    const taskMetadata = {
+      exo__Instance_class: "[[ems__Task]]",
+      ems__Effort_day: "[[2025-10-20]]",
+      ems__Effort_startTimestamp: "2025-10-20T09:00:00",
+      ems__Effort_status: "[[ems__EffortStatusToDo]]",
+      ems__Effort_plannedStartTimestamp: "2025-10-20T10:00:00.000Z",
+      ems__Effort_plannedEndTimestamp: "2025-10-20T12:00:00.000Z",
     };
 
     ctx.mockMetadataExtractor.extractMetadata
