--- packages/obsidian-plugin/tests/unit/renderers/dailyTasks/DailyTasksRenderer.extractFirstValue.test.ts
@@ -1,229 +1,0 @@
-import {
-  setupDailyTasksRendererTest,
-  createMockElement,
-  DailyTasksRendererTestContext,
-  TFile,
-} from "./DailyTasksRenderer.fixtures";
-
-describe("DailyTasksRenderer - extractFirstValue", () => {
-  let ctx: DailyTasksRendererTestContext;
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    ctx = setupDailyTasksRendererTest();
-  });
-
-  it("should extract string value without brackets", async () => {
-    const mockFile = {
-      path: "test.md",
-      parent: { path: "DailyNotes" },
-      basename: "2025-10-20",
-    } as TFile;
-    const metadata = {
-      exo__Instance_class: "[[pn__DailyNote]]",
-      pn__DailyNote_day: "[[2025-10-20]]",
-    };
-
-    const taskFile = {
-      path: "task.md",
-      basename: "task",
-    } as TFile;
-
-    const taskMetadata = {
-      exo__Instance_class: "[[ems__Task]]",
-      ems__Effort_day: "[[2025-10-20]]",
-      ems__Effort_startTimestamp: "2025-10-20T09:00:00",
-      ems__Effort_status: "[[ems__EffortStatusBacklog]]",
-      ems__Effort_area: "[[Development]]",
-    };
-
-    ctx.mockMetadataExtractor.extractMetadata
-      .mockReturnValueOnce(metadata)
-      .mockReturnValueOnce(taskMetadata);
-
-    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
-      "[[pn__DailyNote]]",
-    );
-
-    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
-
-    const mockEl = createMockElement();
-    await ctx.renderer.render(mockEl, mockFile);
-
-    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
-    const getEffortArea = renderCall[1].props.getEffortArea;
-    const result = getEffortArea(taskMetadata);
-    expect(result).toBe("Development");
-  });
-
-  it("should extract first value from array", async () => {
-    const mockFile = {
-      path: "test.md",
-      parent: { path: "DailyNotes" },
-      basename: "2025-10-20",
-    } as TFile;
-    const metadata = {
-      exo__Instance_class: "[[pn__DailyNote]]",
-      pn__DailyNote_day: "[[2025-10-20]]",
-    };
-
-    const taskFile = {
-      path: "task.md",
-      basename: "task",
-    } as TFile;
-
-    const taskMetadata = {
-      exo__Instance_class: "[[ems__Task]]",
-      ems__Effort_day: "[[2025-10-20]]",
-      ems__Effort_startTimestamp: "2025-10-20T09:00:00",
-      ems__Effort_status: "[[ems__EffortStatusBacklog]]",
-      ems__Effort_area: ["[[Development]]", "[[Backend]]"],
-    };
-
-    ctx.mockMetadataExtractor.extractMetadata
-      .mockReturnValueOnce(metadata)
-      .mockReturnValueOnce(taskMetadata);
-
-    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
-      "[[pn__DailyNote]]",
-    );
-
-    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
-
-    const mockEl = createMockElement();
-    await ctx.renderer.render(mockEl, mockFile);
-
-    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
-    const getEffortArea = renderCall[1].props.getEffortArea;
-    const result = getEffortArea(taskMetadata);
-    expect(result).toBe("Development");
-  });
-
-  it("should return null for empty value", async () => {
-    const mockFile = {
-      path: "test.md",
-      parent: { path: "DailyNotes" },
-      basename: "2025-10-20",
-    } as TFile;
-    const metadata = {
-      exo__Instance_class: "[[pn__DailyNote]]",
-      pn__DailyNote_day: "[[2025-10-20]]",
-    };
-
-    const taskFile = {
-      path: "task.md",
-      basename: "task",
-    } as TFile;
-
-    const taskMetadata = {
-      exo__Instance_class: "[[ems__Task]]",
-      ems__Effort_day: "[[2025-10-20]]",
-      ems__Effort_startTimestamp: "2025-10-20T09:00:00",
-      ems__Effort_status: "[[ems__EffortStatusBacklog]]",
-    };
-
-    ctx.mockMetadataExtractor.extractMetadata
-      .mockReturnValueOnce(metadata)
-      .mockReturnValueOnce(taskMetadata);
-
-    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
-      "[[pn__DailyNote]]",
-    );
-
-    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
-
-    const mockEl = createMockElement();
-    await ctx.renderer.render(mockEl, mockFile);
-
-    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
-    const getEffortArea = renderCall[1].props.getEffortArea;
-    const result = getEffortArea(taskMetadata);
-    expect(result).toBeNull();
-  });
-
-  it("should return null for empty array", async () => {
-    const mockFile = {
-      path: "test.md",
-      parent: { path: "DailyNotes" },
-      basename: "2025-10-20",
-    } as TFile;
-    const metadata = {
-      exo__Instance_class: "[[pn__DailyNote]]",
-      pn__DailyNote_day: "[[2025-10-20]]",
-    };
-
-    const taskFile = {
-      path: "task.md",
-      basename: "task",
-    } as TFile;
-
-    const taskMetadata = {
-      exo__Instance_class: "[[ems__Task]]",
-      ems__Effort_day: "[[2025-10-20]]",
-      ems__Effort_startTimestamp: "2025-10-20T09:00:00",
-      ems__Effort_status: "[[ems__EffortStatusBacklog]]",
-      ems__Effort_area: [],
-    };
-
-    ctx.mockMetadataExtractor.extractMetadata
-      .mockReturnValueOnce(metadata)
-      .mockReturnValueOnce(taskMetadata);
-
-    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
-      "[[pn__DailyNote]]",
-    );
-
-    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
-
-    const mockEl = createMockElement();
-    await ctx.renderer.render(mockEl, mockFile);
-
-    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
-    const getEffortArea = renderCall[1].props.getEffortArea;
-    const result = getEffortArea(taskMetadata);
-    expect(result).toBeNull();
-  });
-
-  it("should return null for whitespace-only string", async () => {
-    const mockFile = {
-      path: "test.md",
-      parent: { path: "DailyNotes" },
-      basename: "2025-10-20",
-    } as TFile;
-    const metadata = {
-      exo__Instance_class: "[[pn__DailyNote]]",
-      pn__DailyNote_day: "[[2025-10-20]]",
-    };
-
-    const taskFile = {
-      path: "task.md",
-      basename: "task",
-    } as TFile;
-
-    const taskMetadata = {
-      exo__Instance_class: "[[ems__Task]]",
-      ems__Effort_day: "[[2025-10-20]]",
-      ems__Effort_startTimestamp: "2025-10-20T09:00:00",
-      ems__Effort_status: "[[ems__EffortStatusBacklog]]",
-      ems__Effort_area: "   ",
-    };
-
-    ctx.mockMetadataExtractor.extractMetadata
-      .mockReturnValueOnce(metadata)
-      .mockReturnValueOnce(taskMetadata);
-
-    ctx.mockMetadataExtractor.extractInstanceClass.mockReturnValueOnce(
-      "[[pn__DailyNote]]",
-    );
-
-    ctx.mockVaultAdapter.getAllFiles.mockReturnValue([taskFile]);
-
-    const mockEl = createMockElement();
-    await ctx.renderer.render(mockEl, mockFile);
-
-    const renderCall = ctx.mockReactRenderer.render.mock.calls[0];
-    const getEffortArea = renderCall[1].props.getEffortArea;
-    const result = getEffortArea(taskMetadata);
-    expect(result).toBeNull();
-  });
-});
